refactor: 패키지 분리에 따른 클래스 이동 및 미사용 클래스 삭제(move: JwtAuthConfig, delete: BearerAuthorizationExtractor)
