--- src/main/java/roomescape/authentication/BearerAuthorizationExtractor.java
@@ -1,28 +1,0 @@
-package roomescape.authentication;
-
-import jakarta.servlet.http.HttpServletRequest;
-
-import java.util.Enumeration;
-
-public class BearerAuthorizationExtractor implements AuthorizationExtractor<String> {
-    private static final String BEARER_TYPE = "Bearer";
-    private static final String ACCESS_TOKEN_TYPE = BearerAuthorizationExtractor.class.getSimpleName() + ".ACCESS_TOKEN_TYPE";
-
-    @Override
-    public String extract(HttpServletRequest request) {
-        Enumeration<String> headers = request.getHeaders(AUTHORIZATION);
-        while (headers.hasMoreElements()) {
-            String value = headers.nextElement();
-            if ((value.toLowerCase().startsWith(BEARER_TYPE.toLowerCase()))) {
-                String authHeaderValue = value.substring(BEARER_TYPE.length()).trim();
-                request.setAttribute(ACCESS_TOKEN_TYPE, value.substring(0, BEARER_TYPE.length()).trim());
-                int commaIndex = authHeaderValue.indexOf(',');
-                if (commaIndex > 0) {
-                    authHeaderValue = authHeaderValue.substring(0, commaIndex);
-                }
-                return authHeaderValue;
-            }
-        }
-        return null;
-    }
-}
